refactor: use typescript
Use typescript to improve type safety
