+++ package.json
@@ -68,6 +68,7 @@
     "prop-types": "^15.8.1",
     "semantic-release": "^23.0.4",
     "style-loader": "^3.3.2",
+    "typescript": "^5.4.2",
     "url-loader": "^4.1.1",
     "vite": "^5.1.5"
   }
