--- src/chat/service/index.mjs
@@ -1,1 +1,0 @@
-export * from "./openai.mjs";
