+++ src/index.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import * as ReactDOMClient from "react-dom/client";
-import ChatApp from "./chat/ChatApp.jsx";
-import "./i18n/config.js";
+import ChatApp from "./chat/ChatApp.tsx";
+import "./i18n/config.ts";
 
 const root = ReactDOMClient.createRoot(document.getElementById("app"));
 root.render(
